position and height values should definitely be floats

+++ lib/demoloop-lib/include/helpers.h
@@ -11,8 +11,8 @@
 #include "hsl.h"
 
 GLuint loadTexture(const std::string &path);
-void renderTexture(GLuint texture, int x, int y, int w, int h);
-void renderTexture(GLuint texture, int x, int y, int z, int w, int h);
+void renderTexture(GLuint texture, float x, float y, float w, float h);
+void renderTexture(GLuint texture, float x, float y, float z, float w, float h);
 
 /*
  * Log an SDL error with some error message to the output stream of our choice
